Moved scripts into separated folder
